refactor: abstract campusDataQuery away from App.tsx and move JSON types to context/campusData
