+++ matrufsc/src/App.tsx
@@ -1,14 +1,12 @@
 import { createEffect, createResource, createSignal } from "solid-js";
 import { makePersisted } from "@solid-primitives/storage";
-import {
-    getDisciplinaFromJSON,
-    type JSONCampus,
-    type JSONCampusCode,
-    type JSONDisciplina,
-} from "~/context/plano/parser";
+import { createPersistedSignal } from "~/lib/createPersistedSignal";
+import { createCachedResource } from "~/lib/createCachedResource";
+// Context
+import { campusDataQuery, type JSONCampusCode } from "~/lib/campusDataQuery";
 import { usePlano } from "~/context/plano/Plano.store";
 import { MateriaExistsError } from "~/context/plano/errors";
-import { createPersistedSignal } from "./lib/createPersistedSignal";
+import { getDisciplinaFromJSON, type JSONDisciplina } from "~/context/plano/parser";
 // Components
 import Header from "~/components/header/Header";
 import Footer from "~/components/footer/Footer";
@@ -18,7 +16,6 @@ import Materias from "~/components/materias/Materias";
 import Turmas from "~/components/turmas/Turmas";
 import Horarios from "~/components/horarios/Horarios";
 import CombinacaoSpinner from "./components/horarios/CombinacaoSpinner";
-import { createCachedQuery, createCachedResource } from "./lib/createCachedResource";
 
 const CAMPUS: { title: string; value: JSONCampusCode }[] = [
     { title: "Florianópolis", value: "FLO" },
@@ -27,14 +24,6 @@ const CAMPUS: { title: string; value: JSONCampusCode }[] = [
     { title: "Araranguá", value: "ARA" },
     { title: "Blumenau", value: "BLN" },
 ];
-
-type CampusDataSource = { semester: string; campus: JSONCampusCode };
-
-function campusDataKey(source: CampusDataSource) {
-    return `matrufsc:campusData:${source.campus}_${source.semester}`;
-}
-
-const campusDataQuery = createCachedQuery(fetchCampusData, { key: campusDataKey });
 
 export default function App() {
     const { addMateria, materias } = usePlano();
@@ -134,20 +123,6 @@ async function fetchAvailableSemesters(): Promise<string[]> {
         return json.semesters?.sort().reverse() ?? [];
     } catch (error) {
         console.error("Error fetching semesters:", error);
-        throw error;
-    }
-}
-
-async function fetchCampusData(source: CampusDataSource): Promise<JSONCampus> {
-    const { semester, campus } = source;
-    const ano = semester.slice(0, 4);
-
-    try {
-        const resp = await fetch(`${import.meta.env.BASE_URL}data/${ano}/${semester}-${campus}.json`);
-        if (!resp.ok) throw new Error(`Failed to load campus data: ${resp.status} ${resp.statusText}`);
-        return await resp.json();
-    } catch (error) {
-        console.error("Error fetching campus data:", error);
         throw error;
     }
 }
